utils file created
Created utils file to make app.jsx a bit less busy
